Fix: Move ReaderPostDetailUiStateBuilder outside viewmodels package
